created admin dashboard and debugging

+++ database/migrations/2025_11_21_223644_add_phone_to_donations_table.php
@@ -8,15 +8,20 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::table('donations', function (Blueprint $table) {
+      Schema::table('donations', function (Blueprint $table) {
+        if (!Schema::hasColumn('donations', 'phone')) {
             $table->string('phone')->nullable()->after('payment_method');
-        });
+        }
+    });  
     }
 
     public function down()
     {
-        Schema::table('donations', function (Blueprint $table) {
+       Schema::table('donations', function (Blueprint $table) {
+        if (Schema::hasColumn('donations', 'phone')) {
             $table->dropColumn('phone');
-        });
+        }
+    });
+    
     }
 };
